+++ examples/test_failing_downgrade/conftest.py
@@ -1,0 +1,3 @@
+from pytest_mock_resources import create_postgres_fixture
+
+alembic_engine = create_postgres_fixture()