+++ examples/test_minimum_downgrade_revision/migrations/versions/dddddddddddd_create_foo.py
@@ -1,0 +1,15 @@
+import sqlalchemy as sa
+from alembic import op
+
+revision = "dddddddddddd"
+down_revision = "cccccccccccc"
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    op.add_column("foo", sa.Column("bar_id", sa.Integer(), server_default="9"))
+
+
+def downgrade():
+    op.drop_column("foo", "bar_id")