+++ examples/test_minimum_downgrade_revision/models.py
@@ -1,0 +1,13 @@
+import sqlalchemy
+from sqlalchemy import Column, types
+from sqlalchemy.ext.declarative import declarative_base
+
+Base = declarative_base()
+
+
+class CreatedAt(Base):
+    __tablename__ = "foo"
+
+    id = Column(types.Integer(), autoincrement=True, primary_key=True)
+    foo_id = Column(types.Integer())
+    bar_id = Column(types.Integer())