+++ examples/test_unimplemented_downgrade_warning/conftest.py
@@ -1,0 +1,4 @@
+import pytest
+from pytest_mock_resources import create_postgres_fixture
+
+alembic_engine = create_postgres_fixture()